requirements.txt
discord.py>=2.3.0
aiohttp>=3.8.0
aiosqlite>=0.19.0
python-dateutil>=2.8.0
colorama>=0.4.6
pytz>=2023.3